% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Correlation tests}
\usage{
cor_test(var1, var2, method = NULL, round = TRUE)
}
\arguments{
\item{var1}{a vector to correlate with var2}

\item{var2}{a vector to correlate with var1}

\item{method}{the correlation method to use, either pearson, spearman, or
kendall}

\item{round}{logical, whether or not to round the results}
}
\value{
Returns a data.frame version of the standard htest output. Use
  intended for presentation of data because numbers are converted to
  character and then back to numeric, trimming each values length and
  costing you precision.
}
\description{
A wrapper for cor.test that returns data in a data.frame rather than a
cumbersome list
}
\examples{
cor_test(sample(1:100, 100), sample(1:100, 100), method = "pearson")

}

