% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list.R
\name{create_list}
\alias{create_list}
\title{Create a list of named lists}
\usage{
create_list(list.names, list.lengths)
}
\arguments{
\item{list.names}{a string vector containing list names}

\item{list.lengths}{a numeric value that determines the length of each \code{list.names} list}
}
\value{
Returns a list of named lists, each of length \code{list.lengths}.
}
\description{
Use \code{create_list} to create a list of named lists, each of length \code{list.lengths}.
}
\examples{
# Generate lists, each of length 4
create_list(c("a", "b", "c"), 4)
}

