% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_table.R
\name{data_frame_table}
\alias{data_frame_table}
\title{Create a data.frame table}
\usage{
data_frame_table(data1, prop = TRUE, perc = TRUE)
}
\arguments{
\item{data1}{a vector or data.frame column}

\item{prop}{logical, if \code{TRUE} returns an additional proportion column}
}
\value{
a data.frame table with an optional proportion column
}
\description{
Create a data.frame table with an optional proportion column
}
\examples{
data_frame_table(iris2$Species)
data_frame_table(iris2$Species, prop = FALSE)
}

