% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvariables.R
\name{dclass}
\alias{dclass}
\title{Data.frame of Variables and Classes}
\usage{
dclass(data1, as.list = FALSE)
}
\arguments{
\item{data1}{a \code{data.frame}}

\item{as.list}{logical, \code{TRUE} returns a list split by class}
}
\value{
Returns a data.frame or list with the variable names and their
  respective classes. Returns a data.frame by default.
}
\description{
Helper function for \code{\link{dtable}}. Returns information about variable
class in a \code{data.frame}, or \code{list} if \code{as.list = TRUE}.
}
\examples{
# Load sample data
dclass(iris2)

# Return variable and class data in data.frame
dclass(iris2)

# Return variable and class data in list
dclass(iris2, as.list = TRUE)
}
\seealso{
\code{\link{class}} to examine method for extracting class.
}

