% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{detect_class}
\alias{detect_class}
\title{Detect Variable Classes}
\usage{
detect_class(data1, vnames)
}
\arguments{
\item{data1}{a \code{data.frame}}

\item{vnames}{a vector of variable names from \code{data1} to classify}
}
\value{
Returns a list with a vector containing factor variable names, and a
  vector containing numeric variable names.
}
\description{
Helper function for \code{\link{dtable}}. Obtains column classes and splits
variable names into a list of two vectors based on numeric/integer or factor
class.
}
\examples{
# Load sample data
data(iris2)

# All variables
detect_class(iris2, names(iris2))

# Single variable
detect_class(iris2, "Species")
}
\seealso{
\code{\link{class}} to examine method for extracting class type.
}

