% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{detectClass}
\alias{detectClass}
\title{detectClass}
\usage{
detectClass(data, vnames)
}
\arguments{
\item{data}{data.frame}

\item{vnames}{vector of variable names from data.frame to classify}
}
\value{
Returns a list with a vector containing factor variable names, 
   and a vector containing numeric variable names.
}
\description{
Helper function for dtable. Obtains column classes and splits variable 
names into a list of two vectors based on numeric/integer or factor class.
}
\examples{
data(demoData)
detectClass(myData, "Education")

vars <- c("Education", "Age")
detectClass(myData, vars)

}
\seealso{
\code{\link{class}} which this function relies on
}

