% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dnumeric}
\alias{dnumeric}
\title{dnumeric}
\usage{
dnumeric(data, vnames, round = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{vnames}{single variable name to examine with \code{psych::describe}}

\item{round}{logical, returns rounded values if \code{TRUE}}
}
\value{
data.frame
}
\description{
This function converts a numeric column from a data.frame into a descriptive statistics
table using \code{psych::describe}.
}

