% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dtable}
\alias{dtable}
\title{dtable}
\usage{
dtable(data, vnames, round = F, neat = F)
}
\arguments{
\item{data}{data.frame}

\item{vnames}{vector of variable names from data.frame to use in creation
of demographic tables}

\item{round}{logical, TRUE returns rounded numeric table}

\item{neat}{logical, TRUE returns rounded factor table with percent symbols}
}
\value{
list of data.frames $factor and $numeric if `vnames` contains both classes,
   single data.frame returned if only one variable class in vnames.
}
\description{
dtable
}

