% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{extract}
\alias{extract}
\title{Extract and paste from a data.frame}
\usage{
extract(data1, selectCol = NULL, selectRows = NULL, column = NULL)
}
\arguments{
\item{data1}{a \code{data.frame}}

\item{selectCol}{select a column}

\item{selectRows}{select value in that column}

\item{column}{select column to extract}
}
\value{
String vector with selected values
}
\description{
Extract and paste from a data.frame
}

