% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvariable.R
\name{factor_length}
\alias{factor_length}
\title{Return factor level lengths for all variables in a data.frame}
\usage{
factor_length(data1)
}
\arguments{
\item{data1}{a \code{data.frame}}
}
\value{
a \code{data.frame} with variable names and factor level lengths
}
\description{
Return factor level lengths for all variables in a data.frame
}
\examples{
factor_length(iris2)
}

