% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.R
\name{fe}
\alias{fe}
\title{file.edit 2.0}
\usage{
fe(rowNum = 1)
}
\arguments{
\item{rowNum}{integer, specify row number from `lf()` to open it's
corresponding file}
}
\value{
nothing, opens the file specified
}
\description{
The usage of `file.edit`, but easier.
}
\examples{
# Open the first file in your current wd
fe(1)

}

