% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{lf}
\alias{lf}
\title{list.files 2.0}
\usage{
lf(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{string, path to list files from}

\item{recursive}{logical, list files within folders}
}
\value{
a one column data.frame named after the working directory that
  contains a recursive file list of that directory's contents as rows
}
\description{
The output of `list.files`, but easier to read.
}
\examples{
lf()

}

