% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dfactor}
\alias{dfactor}
\title{dfactor}
\usage{
dfactor(data, vnames, neat = FALSE, sizesort = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{vnames}{vector of variable names from data.frame to use in creation
of demographics table, multiple names examines crosstabs.}

\item{neat}{logical, TRUE returns rounded factor table with percent symbols}

\item{sizesort}{logical, TRUE returns table sorted by size}
}
\value{
data.frame
}
\description{
dfactor
}

