% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dfactor}
\alias{dfactor}
\title{dfactor}
\usage{
dfactor(data, vnames, neat = TRUE, sizesort = TRUE)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{vnames}{vector of variable names from \code{data.frame} to use in creation
of demographics table, multiple names produces crosstabs.}

\item{neat}{logical, \code{TRUE} returns rounded factor table with percent symbols}

\item{sizesort}{logical, \code{TRUE} returns table sorted by size}
}
\value{
data.frame
}
\description{
This function converts a column(s) from a data.frame into a frequencies table
formatted in standard presentation structure with percent symbols.
}

