% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R, R/dfactor.R
\name{dfactor}
\alias{dfactor}
\title{Demographic Factor Frequencies Tables}
\usage{
dfactor(data1, vnames, neat = TRUE, sizesort = TRUE)

dfactor(data1, vnames, neat = TRUE, sizesort = TRUE)
}
\arguments{
\item{vnames}{a vector of variable names from \code{data.frame} to use in
creation of demographic frequency table, multiple variables produces more
specific groups.}

\item{neat}{logical, \code{TRUE} returns rounded factor table with percent
symbols}

\item{sizesort}{logical, \code{TRUE} returns table sorted by size}

\item{data}{a \code{data.frame}}

\item{data}{a \code{data.frame}}

\item{vnames}{a vector of variable names from \code{data.frame} to use in
creation of demographic frequency table, multiple variables produces more
specific groups.}

\item{neat}{logical, \code{TRUE} returns rounded factor table with percent
symbols}

\item{sizesort}{logical, \code{TRUE} returns table sorted by size}
}
\value{
Returns a demographic frequency table of varying specificity.

Returns a demographic frequency table of varying specificity.
}
\description{
This function converts a column(s) from a \code{data.frame} into a
frequencies table formatted in standard presentation structure with percent
symbols.

This function converts a column(s) from a \code{data.frame} into a
frequencies table formatted in standard presentation structure with percent
symbols.
}
\examples{
# Single demographic
dfactor(iris2, "Species")

# Two demographics
dfactor(iris2, c("Color", "Species"))
# Single demographic
dfactor(iris2, "Species")

# Two demographics
dfactor(iris2, c("Color", "Species"))
}
\seealso{
\link{\code{table}} which this function utilizes.

\code{\link{table}} which this function utilizes.
}

