% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_table.R
\name{dft}
\alias{dft}
\title{Create a data.frame table}
\usage{
dft(data1, prop = TRUE, perc = TRUE)
}
\arguments{
\item{data1}{a vector or data.frame column}

\item{prop}{logical, if \code{TRUE} returns an additional proportion column}

\item{perc}{logical, if \code{TRUE} returns an additional percentage column}
}
\value{
a data.frame table with optional proportion and percentage columns
}
\description{
Create a table with a data.frame structure and optional proportion and
percentage columns
}
\examples{
data_frame_table(iris2$Species)
data_frame_table(iris2$Species, prop = FALSE)
}

