% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dnumeric}
\alias{dnumeric}
\title{dnumeric}
\usage{
dnumeric(data, vnames, round = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{vnames}{single variable name to describe}

\item{round}{}
}
\value{
data.frame
}
\description{
dnumeric
}

