% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dtable}
\alias{dtable}
\title{dtable}
\usage{
dtable(data, vnames, neat = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{vnames}{a vector of variable names from \code{data.frame} to use in creation of
demographic tables}

\item{neat}{logical, \code{TRUE} returns rounded factor table with percent
symbols}
}
\value{
List of two \code{data.frame}s, split into factor and numeric variables if
  \code{vnames} contains both classes, single \code{data.frame} returned if only one
  variable class detected in \code{vnames}.
}
\description{
Demographic frequency tables and descriptive statistics tables for R.
}
\examples{
# Examine all variables
dtable(iris2, names(iris2))

# Examine single variable
dtable(iris2, "Species")

# Round all output and add percent symbols to factor output
dtable(iris2, names(iris2), neat = TRUE)
}

