% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dtable}
\alias{dtable}
\title{dtable}
\usage{
dtable(data, vnames, round = TRUE, neat = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{vnames}{vector of variable names from data.frame to use in creation
of demographic tables}

\item{round}{logical, \code{TRUE} returns rounded numeric table}

\item{neat}{logical, \code{TRUE} returns rounded factor table with percent symbols}
}
\value{
list of data.frames $factor and $numeric if `vnames` contains both classes,
   single data.frame returned if only one variable class in vnames.
}
\description{
Demographic frequency tables and descriptive statistics tables for R.
}
\examples{
dtable(iris, names(iris))

}

