% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R
\name{dtable}
\alias{dtable}
\title{Generate demographic frequencies and descriptive statistics tables}
\usage{
dtable(data1, variables = NULL, frequencies = NULL, descriptives = NULL,
  neat = TRUE, as.list = FALSE, sizesort = TRUE)
}
\arguments{
\item{neat}{logical, defaults to \code{TRUE} and returns rounded factor table
with percent symbols}

\item{sizesort}{logical, returns data sorted by frequency and mean if
\code{TRUE}}

\item{data}{a \code{data.frame}}

\item{vnames}{all, or a subset \code{vector} of variable names from
\code{data.frame}}
}
\value{
List of two \code{data.frames}, split into \code{factor} and
  \code{numeric} variables if \code{vnames} contains both classes, single
  \code{data.frame} returned if only one variable class detected in
  \code{vnames} by \code{dclass}.
}
\description{
Generate demographic frequencies tables and descriptive statistics tables in
data.frame format for R.
}
\examples{
# Examine all variables
dtable(iris2)

# Examine single variable - Out of service
dtable(iris2, "Species")

# Examine two or more variables  - Out of service
dtable(iris2, c("Attractiveness", "Species"))

# Raw output
dtable(iris2, neat = FALSE)

}
\seealso{
\code{\link{dclass}} to see how class is identified.
}

