% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotables.R, R/dtable.R
\name{dtable}
\alias{dtable}
\title{Generate demographic frequencies and descriptive statistics tables}
\usage{
dtable(data1, variables = NULL, by = NULL, frequencies = NULL,
  statistics = NULL, neat = TRUE, as.list = FALSE, sizesort = FALSE)

dtable(data1, variables = NULL, by = NULL, frequencies = NULL,
  statistics = NULL, neat = TRUE, as.list = FALSE, sizesort = FALSE)
}
\arguments{
\item{data1}{a \code{data.frame} or \code{matrix}}

\item{variables}{select which columns of \code{data1} to analyze}

\item{frequencies}{select which columns to create frequencies tables for}

\item{statistics}{select which columns to create statistics tables for}

\item{neat}{logical, defaults to \code{TRUE} and returns rounded factor table
with percent symbols}

\item{as.list}{logical, if \code{TRUE} it returns frequencies and statistics
tables in a list format, split by \code{variables}}

\item{sizesort}{logical, returns data sorted by frequency and mean if
\code{TRUE}}

\item{data}{a \code{data.frame}}

\item{vnames}{all, or a subset \code{vector} of variable names from
\code{data.frame}}

\item{neat}{logical, if \code{TRUE} returns rounded and formatted tables}

\item{sizesort}{logical, if \code{TRUE} returns data sorted by frequency and
mean}
}
\value{
List of two \code{data.frames}, split into \code{factor} and
 \code{numeric} variables if \code{vnames} contains both classes, single
 \code{data.frame} returned if only one variable class detected in
 \code{vnames} by \code{dclass}.

Returns descriptive frequencies and statistics tables for all
  variables in \code{data1} by default. Predicts whether to create a
  frequencies table, statistics table, or both for each varible based on
  information gathered using \link{\code{dvariable}}.
}
\description{
Simplifies the process and reduces the amount of code involved in generating
descriptive frequencies and statistics tables with or without formatting.

Simplifies the process and reduces the amount of code involved in generating
descriptive frequencies and statistics tables. Output can be generated with
or without formatting and rounding, and either split into lists by variable,
or aggregated into two data.frames.
}
\examples{
# Examine all variables
dtable(iris2)

# Examine single variable - Out of service
dtable(iris2, "Species")

# Examine two or more variables  - Out of service
dtable(iris2, c("Attractiveness", "Species"))

# Raw output
dtable(iris2, neat = FALSE)

# Analyze all data
dtable(iris2)

#Analyze two or more variables
dtable(iris2, variables = c("Color", "Sold", "LikelyToBuy"))

# Analyze a single variable
dtable(iris2, variables = "Color")

# Return raw output
dtable(iris2, neat = FALSE)

# Return list output
dtable(iris2, as.list = TRUE)

# Frequencies sorted by size
dtable(iris2, sizesort = TRUE)

}
\seealso{
\code{\link{dclass}} to see how class is identified.
}

